backend/services/inventory_srv/requirements.txt
redis
peewee
pymysql
passlib
grpcio
grpcio-tools==1.43.0
protobuf
loguru
python-consul2
rocketmq-client-python
nacos-sdk-python